riscv/cfi: Add tunables for controlling CFI extensions
Two tunables, glibc.cpu.riscv_cfi_lp and glibc.cpu.riscv_cfi_ss are
added for overriding the behavior of CFI features.

There are three options available for each, which are on|off|permissive.
"on" will force a feature to always stays turned on, and vice versa for
"off". Use "permissive" if you want to allow the executable to
dynamically load a legacy object (not compiled with CFI supports).

Only on and off is implemented so far.